requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
sentence-transformers
faiss-cpu
pypdf
python-dotenv
streamlit